Compute Relative Expression Levels

This function computes the relative expression profiles.

In detail, the Relative Expression Profile is being computed as follows over developmental stages s:

$$f_s = (e_s - e_min)/(e_max - e_min)$$,

as follows over cells:

$$f_c = (e_c - e_min)/(e_max - e_min)$$,

as follows over celltypes:

$$f_ct = (e_ct - e_min)/(e_max - e_min)$$,

as follows over phylostrata:

$$f_ps = (e_ps - e_min)/(e_max - e_min)$$,

where e_min and e_max denote either the minimum/maximum mean expression level over developmental stages \(s\), cells \(c\), celltypes \(ct\) or phylostrata \(ps\).

This linear transformation corresponds to a shift by \(e_min - e_max\). As a result, the relative expression level \(f_s\) of developmental stage \(s\), \(f_c\) of cell \(c\), \(f_ct\) of celltype \(ct\) or \(f_ps\) of phylotstratum \(ps\) with minimum \(e_s\), \(e_c\), \(e_ct\) or \(e_ps\) is 0, whereas the relative expression level \(f_s\) of developmental stage \(s\), \(f_c\) of cell \(c\), \(f_ct\) of celltype \(ct \) or \(f_ps\) of phylotstratum \(ps\) with maximum \(e_s\), \(e_c\), \(e_ct\) or \(e_ps\) is 1, and the relative expression levels of all other stages \(s\), cells \(c\), celltypes \(ct\) or phylostrata \(ps\) range between 0 and 1.

REMatrix(
  ExpressionSet,
  Phylostratum = NULL,
  by = NULL,
  groups = NULL,
  split = 1e+05,
  showprogress = TRUE,
  threads = 1
)

Arguments

ExpressionSet

expression object with rownames as GeneID (dgCMatrix) or standard PhyloExpressionSet object.

Phylostratum

a named vector representing phylostratum per GeneID with names as GeneID (not used if Expression is PhyloExpressionSet).

by

specify min/max transformation by row (stages, cells, celltypes) or by column (phylostratum)

groups

specify stages or cells to be grouped into celltypes by a named list

split

specify number of columns to split

showprogress

boolean if progressbar should be shown

threads

specify number of threads

Value

a numeric matrix storing the summed partial TEI values for each strata.

Details

The partial TEI values combined per strata give an overall impression of the contribution of each strata to the global TEI pattern.

References

Domazet-Loso T. and Tautz D. (2010). A phylogenetically based transcriptome age index mirrors ontogenetic divergence patterns. Nature (468): 815-818.

Quint M et al. (2012). A transcriptomic hourglass in plant embryogenesis. Nature (490): 98-101.

Drost HG et al. (2015) Mol Biol Evol. 32 (5): 1221-1231 doi:10.1093/molbev/msv012

Author

Kristian K Ullrich

Examples

# reading a standard PhyloExpressionSet
data(PhyloExpressionSetExample, package="myTAI")

# get mean expression
REMatrix(PhyloExpressionSetExample)
#> 12 x 7 sparse Matrix of class "dgCMatrix"
#>      Zygote Quadrant Globular    Heart  Torpedo     Bent   Mature
#> 1  2607.882 2579.372 2604.856 2525.704 2554.825 2622.757 2696.331
#> 2  2597.258 2574.745 2467.679 2388.045 2296.410 2243.716 2321.709
#> 3  2528.272 2363.159 2019.436 2099.079 2155.642 2196.875 2855.866
#> 4  1925.320 1887.078 1771.399 1787.175 1740.823 1867.981 2358.893
#> 5  2378.883 2368.593 2061.729 2077.087 2076.693 2564.904 3157.761
#> 6  1658.253 1697.242 1485.401 1462.613 1492.861 1631.741 2304.683
#> 7  1993.321 1717.659 1480.525 1590.009 1545.078 1600.264 2385.409
#> 8  1781.653 1670.106 1452.180 1414.052 1359.376 1816.718 2364.070
#> 9  1758.119 1764.748 1708.815 1575.727 1388.920 1687.314 2193.930
#> 10 2414.456 2501.390 2163.810 1938.060 1770.039 1993.032 2127.015
#> 11 1999.163 2071.456 1702.779 1710.290 1662.099 1726.865 2501.443
#> 12 2126.189 2036.804 1896.964 1909.578 1859.485 1995.732 2387.343

# computing relative expression profile over stages
REMatrix(PhyloExpressionSetExample, by="row")
#>       Zygote  Quadrant   Globular      Heart    Torpedo       Bent    Mature
#> 1  0.4816246 0.3145330 0.46389184 0.00000000 0.17067495 0.56880234 1.0000000
#> 2  1.0000000 0.9363209 0.63348381 0.40823711 0.14904726 0.00000000 0.2206063
#> 3  0.6083424 0.4109402 0.00000000 0.09521758 0.16284114 0.21213845 1.0000000
#> 4  0.2985050 0.2366309 0.04946941 0.07499453 0.00000000 0.20573325 1.0000000
#> 5  0.2893657 0.2799777 0.00000000 0.01401191 0.01365328 0.45908792 1.0000000
#> 6  0.2323316 0.2786335 0.02706119 0.00000000 0.03592044 0.20084761 1.0000000
#> 7  0.5666979 0.2620602 0.00000000 0.12099252 0.07133814 0.13232551 1.0000000
#> 8  0.4203039 0.3092784 0.09237036 0.05442042 0.00000000 0.45520558 1.0000000
#> 9  0.4586261 0.4668613 0.39738003 0.23205534 0.00000000 0.37067096 1.0000000
#> 10 0.8811321 1.0000000 0.53841500 0.22974016 0.00000000 0.30490542 0.4881046
#> 11 0.4015809 0.4877111 0.04846721 0.05741594 0.00000000 0.07716367 1.0000000
#> 12 0.5052572 0.3359211 0.07100055 0.09489782 0.00000000 0.25811214 1.0000000

# computing relative expression profile over stages with groups
groups<-setNames(list(c("Zygote","Quadrant"),c("Heart","Torpedo")),
    c("A","B"))
REMatrix(PhyloExpressionSetExample, by="row", groups=groups)
#>    A B
#> 1  1 0
#> 2  1 0
#> 3  1 0
#> 4  1 0
#> 5  1 0
#> 6  1 0
#> 7  1 0
#> 8  1 0
#> 9  1 0
#> 10 1 0
#> 11 1 0
#> 12 1 0

# computing relative expression profile over phylostrata
REMatrix(PhyloExpressionSetExample, by="column")
#>       Zygote   Quadrant   Globular      Heart    Torpedo       Bent     Mature
#> 1  1.0000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 0.55233430
#> 2  0.9888127 0.99491182 0.88099246 0.87616751 0.78383425 0.62929677 0.18888670
#> 3  0.9161674 0.76221182 0.49212104 0.61622477 0.66608047 0.58348684 0.70711045
#> 4  0.2812335 0.23862294 0.27693702 0.33564758 0.31908263 0.26182743 0.22496154
#> 5  0.7588542 0.76818853 0.52881183 0.59644095 0.60003989 0.94341960 1.00000000
#> 6  0.0000000 0.02984381 0.02882044 0.04368391 0.11166065 0.03078443 0.17236822
#> 7  0.3528406 0.05229796 0.02459049 0.15828439 0.15534034 0.00000000 0.25068575
#> 8  0.1299455 0.00000000 0.00000000 0.00000000 0.00000000 0.21169254 0.22998340
#> 9  0.1051629 0.10408593 0.22264242 0.14543638 0.02471319 0.08513457 0.06491903
#> 10 0.7963144 0.91423715 0.61737165 0.47137805 0.34352212 0.38412800 0.00000000
#> 11 0.3589932 0.44140023 0.21740628 0.26648478 0.25322888 0.12381623 0.36325921
#> 12 0.4927571 0.40329013 0.38587022 0.44575652 0.41834413 0.38676820 0.25256281