Compute a Permutation Matrix of Transcriptome Evolutionary Index (TEI)

This function computes the transcriptome evolutionary index (TEI) using permuted strata values.

bootTEI(
  ExpressionSet,
  Phylostratum = NULL,
  permutations = 100,
  split = 1e+05,
  showprogress = TRUE,
  threads = 1
)

Arguments

ExpressionSet

expression object with rownames as GeneID (dgCMatrix) or standard PhyloExpressionSet object.

Phylostratum

a named vector representing phylostratum per GeneID with names as GeneID (not used if Expression is PhyloExpressionSet).

permutations

a numeric value specifying the number of permutations to be performed.

split

specify number of columns to split

showprogress

boolean if progressbar should be shown

threads

specify number of threads

Value

a numeric matrix storing the TEI values based on permuted strata.

Details

The strata values are sampled and the global TEI is calculated N times.

References

Domazet-Loso T. and Tautz D. (2010). A phylogenetically based transcriptome age index mirrors ontogenetic divergence patterns. Nature (468): 815-818.

Quint M et al. (2012). A transcriptomic hourglass in plant embryogenesis. Nature (490): 98-101.

Drost HG et al. (2015) Mol Biol Evol. 32 (5): 1221-1231 doi:10.1093/molbev/msv012

Author

Kristian K Ullrich

Examples

# reading a standard PhyloExpressionSet
data(PhyloExpressionSetExample, package="myTAI")

# computing partial TEI contribution per gene
bootTEI(PhyloExpressionSetExample)
#>          Zygote Quadrant Globular    Heart  Torpedo     Bent   Mature
#>   [1,] 3.592840 3.574739 3.556491 3.565015 3.557455 3.550949 3.573868
#>   [2,] 3.489426 3.505456 3.496359 3.490717 3.494249 3.496670 3.450509
#>   [3,] 3.391971 3.406740 3.349821 3.341251 3.366526 3.328624 3.355684
#>   [4,] 3.448940 3.462513 3.441205 3.447871 3.451333 3.450241 3.555925
#>   [5,] 3.616022 3.637903 3.627258 3.628129 3.602680 3.608769 3.499963
#>   [6,] 3.532448 3.542667 3.534321 3.534554 3.525404 3.506566 3.474138
#>   [7,] 3.521225 3.536477 3.518260 3.517332 3.501621 3.477563 3.549316
#>   [8,] 3.587381 3.570252 3.582187 3.584623 3.557954 3.540633 3.566609
#>   [9,] 3.518545 3.510167 3.521175 3.524444 3.520130 3.519052 3.545788
#>  [10,] 3.509193 3.510088 3.516281 3.519664 3.527706 3.554127 3.580456
#>  [11,] 3.520947 3.524591 3.508999 3.511161 3.508115 3.475274 3.495601
#>  [12,] 3.421528 3.421449 3.432093 3.426250 3.418235 3.432611 3.514224
#>  [13,] 3.537188 3.545827 3.528691 3.533464 3.540126 3.529641 3.587330
#>  [14,] 3.531461 3.512157 3.527024 3.542150 3.542221 3.571407 3.566686
#>  [15,] 3.494230 3.496450 3.511688 3.483237 3.469556 3.436386 3.490617
#>  [16,] 3.504578 3.486505 3.521971 3.535636 3.537698 3.501479 3.496164
#>  [17,] 3.537029 3.528066 3.535231 3.556130 3.563763 3.542659 3.508029
#>  [18,] 3.565015 3.555177 3.562955 3.556618 3.577058 3.544029 3.532906
#>  [19,] 3.465737 3.445194 3.469257 3.473844 3.481518 3.494384 3.506870
#>  [20,] 3.526447 3.516243 3.490022 3.519470 3.532934 3.562262 3.557494
#>  [21,] 3.552795 3.558271 3.539629 3.551420 3.572578 3.602524 3.533881
#>  [22,] 3.493036 3.491746 3.519314 3.500008 3.490814 3.477849 3.430342
#>  [23,] 3.574956 3.555820 3.564436 3.563681 3.555073 3.570018 3.563286
#>  [24,] 3.507697 3.521762 3.552362 3.546089 3.552100 3.569377 3.569325
#>  [25,] 3.494538 3.512973 3.509080 3.521439 3.541150 3.581469 3.520782
#>  [26,] 3.440452 3.457399 3.463093 3.486922 3.489425 3.506259 3.531600
#>  [27,] 3.495582 3.527830 3.544066 3.529149 3.540878 3.522782 3.480987
#>  [28,] 3.520967 3.491445 3.472864 3.480022 3.474699 3.451432 3.527504
#>  [29,] 3.475254 3.446273 3.419546 3.444805 3.453329 3.439520 3.479189
#>  [30,] 3.460402 3.456997 3.479343 3.489270 3.508005 3.495476 3.523680
#>  [31,] 3.608583 3.582099 3.593835 3.607427 3.616297 3.604383 3.566699
#>  [32,] 3.505565 3.490384 3.484971 3.485089 3.464878 3.479061 3.494898
#>  [33,] 3.522614 3.533117 3.497064 3.489641 3.509244 3.551930 3.606996
#>  [34,] 3.485572 3.462939 3.446648 3.448310 3.473130 3.481979 3.481180
#>  [35,] 3.486416 3.479302 3.534658 3.534378 3.536412 3.531813 3.568802
#>  [36,] 3.511439 3.499128 3.468045 3.478455 3.495014 3.522046 3.590250
#>  [37,] 3.440599 3.426481 3.430439 3.443186 3.462809 3.473270 3.465199
#>  [38,] 3.457407 3.471955 3.494976 3.505055 3.527557 3.539825 3.516558
#>  [39,] 3.496186 3.509518 3.487963 3.495311 3.480130 3.486330 3.484697
#>  [40,] 3.493609 3.490185 3.515169 3.510797 3.511999 3.496951 3.518684
#>  [41,] 3.570690 3.567063 3.584103 3.589555 3.575299 3.577228 3.453065
#>  [42,] 3.518575 3.514557 3.543691 3.549589 3.533396 3.523082 3.472383
#>  [43,] 3.428519 3.447546 3.463591 3.466307 3.475290 3.451989 3.433131
#>  [44,] 3.507714 3.485483 3.443337 3.451344 3.470690 3.465174 3.470543
#>  [45,] 3.461622 3.485544 3.486777 3.478984 3.474669 3.434378 3.426921
#>  [46,] 3.454257 3.477336 3.483102 3.481758 3.474805 3.423000 3.422915
#>  [47,] 3.534415 3.525514 3.514775 3.516420 3.517422 3.501309 3.585723
#>  [48,] 3.485761 3.472357 3.467349 3.480850 3.480501 3.450440 3.451998
#>  [49,] 3.581468 3.568190 3.552684 3.557042 3.577126 3.628371 3.527365
#>  [50,] 3.485724 3.471767 3.456267 3.453270 3.468053 3.451176 3.508365
#>  [51,] 3.520255 3.525863 3.520170 3.511316 3.502593 3.487583 3.400506
#>  [52,] 3.562610 3.551992 3.543972 3.556202 3.536455 3.489782 3.535486
#>  [53,] 3.570857 3.524313 3.507236 3.518092 3.509351 3.516950 3.601920
#>  [54,] 3.590550 3.585744 3.606349 3.597213 3.551219 3.516510 3.575735
#>  [55,] 3.451091 3.445061 3.469608 3.487711 3.508204 3.510542 3.560582
#>  [56,] 3.521221 3.515656 3.511329 3.501040 3.482257 3.471601 3.521075
#>  [57,] 3.507034 3.531619 3.508092 3.484910 3.490003 3.480906 3.472237
#>  [58,] 3.517074 3.491980 3.470956 3.490900 3.491091 3.497482 3.481353
#>  [59,] 3.367451 3.362287 3.357935 3.353388 3.366865 3.347860 3.430352
#>  [60,] 3.471374 3.498673 3.502571 3.508157 3.509535 3.459662 3.394326
#>  [61,] 3.496431 3.503243 3.483423 3.468589 3.471413 3.456981 3.510238
#>  [62,] 3.561156 3.524965 3.496800 3.498363 3.513005 3.515499 3.480140
#>  [63,] 3.498783 3.466536 3.464908 3.468227 3.465628 3.488203 3.574985
#>  [64,] 3.478589 3.488641 3.526742 3.530895 3.538197 3.568100 3.518842
#>  [65,] 3.458294 3.477894 3.434255 3.424718 3.446163 3.441752 3.444568
#>  [66,] 3.591181 3.560872 3.571392 3.586372 3.577457 3.568879 3.495106
#>  [67,] 3.358448 3.351669 3.372408 3.373061 3.371327 3.409126 3.432355
#>  [68,] 3.458902 3.435666 3.470500 3.475539 3.488203 3.538812 3.486558
#>  [69,] 3.453136 3.485758 3.498056 3.505387 3.500675 3.485283 3.449634
#>  [70,] 3.404584 3.422377 3.427218 3.442027 3.425526 3.435208 3.464158
#>  [71,] 3.560085 3.557370 3.549340 3.570077 3.566362 3.562751 3.477466
#>  [72,] 3.476220 3.476205 3.481349 3.492050 3.494544 3.510316 3.529459
#>  [73,] 3.596921 3.574025 3.558915 3.550473 3.546775 3.525027 3.518952
#>  [74,] 3.400999 3.405315 3.408929 3.422190 3.434220 3.438935 3.456177
#>  [75,] 3.496554 3.502931 3.536735 3.543888 3.562199 3.586574 3.607049
#>  [76,] 3.490795 3.490250 3.519188 3.523036 3.496753 3.471189 3.565060
#>  [77,] 3.489327 3.474469 3.460312 3.452558 3.465338 3.448885 3.435238
#>  [78,] 3.473735 3.486264 3.490888 3.487814 3.527183 3.527059 3.543595
#>  [79,] 3.465572 3.470588 3.468183 3.466633 3.490353 3.480178 3.411905
#>  [80,] 3.512899 3.495921 3.501752 3.523538 3.512880 3.506844 3.536806
#>  [81,] 3.533117 3.523338 3.531062 3.544762 3.558518 3.571304 3.531464
#>  [82,] 3.556654 3.549949 3.535368 3.532572 3.531800 3.511713 3.494740
#>  [83,] 3.464165 3.473312 3.477694 3.476396 3.462746 3.481965 3.525500
#>  [84,] 3.552897 3.518699 3.514517 3.540418 3.538532 3.556933 3.528756
#>  [85,] 3.558227 3.568154 3.528402 3.528774 3.521285 3.508184 3.491314
#>  [86,] 3.405926 3.407545 3.432306 3.431061 3.446228 3.516896 3.517256
#>  [87,] 3.545167 3.547963 3.595315 3.605737 3.594052 3.600550 3.567354
#>  [88,] 3.571222 3.547879 3.536799 3.523264 3.524317 3.541390 3.536031
#>  [89,] 3.494873 3.498760 3.474525 3.473720 3.473214 3.469810 3.528499
#>  [90,] 3.499968 3.476687 3.455422 3.463884 3.467351 3.493354 3.487649
#>  [91,] 3.533881 3.554603 3.544782 3.526076 3.497927 3.464949 3.452701
#>  [92,] 3.542411 3.554843 3.530515 3.522451 3.514872 3.512520 3.499843
#>  [93,] 3.428723 3.414518 3.410463 3.412962 3.428980 3.436197 3.522383
#>  [94,] 3.549037 3.546293 3.568482 3.571764 3.536232 3.510461 3.494974
#>  [95,] 3.438923 3.445211 3.458776 3.461619 3.484288 3.501693 3.556601
#>  [96,] 3.506651 3.515978 3.556979 3.550836 3.541191 3.509316 3.555685
#>  [97,] 3.479166 3.482444 3.509934 3.503015 3.489801 3.502275 3.540864
#>  [98,] 3.545117 3.542748 3.529227 3.518181 3.484493 3.464586 3.480592
#>  [99,] 3.572879 3.570041 3.584772 3.614148 3.611953 3.585427 3.512966
#> [100,] 3.573490 3.554466 3.527042 3.533702 3.538004 3.514437 3.502118